Restructure ConversionFunnelChart
