+++ projects/js-packages/charts/src/components/conversion-funnel-chart/private/conversion-funnel-chart.module.scss
