+++ projects/js-packages/charts/src/components/conversion-funnel-chart/private/index.ts
@@ -1,0 +1,2 @@
+export { default as styles } from './conversion-funnel-chart.module.scss';
+export { useFunnelSelection } from './use-funnel-selection';
