+++ projects/js-packages/charts/src/components/conversion-funnel-chart/private/use-funnel-selection.ts
