+++ projects/js-packages/charts/src/utils/color-utils.ts
