+++ projects/js-packages/charts/src/utils/test/color-utils.test.ts
@@ -1,4 +1,4 @@
-import { hexToRgba } from '../utils/color-utils';
+import { hexToRgba } from '../color-utils';
 
 describe( 'hexToRgba', () => {
 	describe( 'Valid hex colors', () => {
